requirements.txt
pandas
python-dotenv
oracledb
pyodbc
requests
gspread
google-auth
msal
